+++ include/vrv/devicecontext.h
@@ -147,7 +147,7 @@ public:
         = 0;
     virtual void DrawRectangle(int x, int y, int width, int height) = 0;
     virtual void DrawRotatedText(const std::string &text, int x, int y, double angle) = 0;
-    virtual void DrawRoundedRectangle(int x, int y, int width, int height, double radius) = 0;
+    virtual void DrawRoundedRectangle(int x, int y, int width, int height, int radius) = 0;
     virtual void DrawText(const std::string &text, const std::wstring wtext = L"", int x = VRV_UNSET, int y = VRV_UNSET)
         = 0;
     virtual void DrawMusicText(const std::wstring &text, int x, int y, bool setSmuflGlyph = false) = 0;
@@ -200,7 +200,9 @@ public:
      * For example, the method can be used for grouping shapes in <g></g> in SVG
      */
     ///@{
-    virtual void StartGraphic(Object *object, std::string gClass, std::string gId, bool preprend = false) = 0;
+    virtual void StartGraphic(
+        Object *object, std::string gClass, std::string gId, bool primary = true, bool preprend = false)
+        = 0;
     virtual void EndGraphic(Object *object, View *view) = 0;
     ///@}