+++ src/devicecontext.cpp
@@ -202,7 +202,10 @@ void DeviceContext::AddGlyphToTextExtend(Glyph *glyph, TextExtend *extend)
     tmp = advX * m_fontStack.top()->GetPointSize();
     advX = ceil(tmp / (double)glyph->GetUnitsPerEm());
 
-    extend->m_width += std::max(partialWidth + x, advX);
+    // Changed because the width should only be the sum of advX
+    // Alternatively we could add what is below 0 for the first and what is beyond the advx for the last
+    //extend->m_width += std::max(partialWidth + x, advX);
+    extend->m_width += advX;
     extend->m_height = std::max(partialHeight, extend->m_height);
     extend->m_ascent = std::max(partialHeight + y, extend->m_ascent);
     extend->m_descent = std::max(-y, extend->m_descent);