+++ src/floatingobject.cpp
@@ -54,7 +54,7 @@ FloatingObject::FloatingObject() : Object("fe")
     Reset();
 }
 
-FloatingObject::FloatingObject(std::string classid) : Object(classid)
+FloatingObject::FloatingObject(const std::string &classid) : Object(classid)
 {
     Reset();