+++ travis/ci_precache.sh
@@ -64,7 +64,7 @@ cmake ../cmake
 make -j 8
 
 echo "Update the documentation of the option list"
-./verovio -? > $OUTPUT_DIRECTORY/_includes/cli.txt
+./verovio -h > $OUTPUT_DIRECTORY/_includes/cli.txt
 
 # Return to the root
 cd ..